+++ apps/example/src/components/peaks/PeaksList.jsx
@@ -12,11 +12,11 @@ import {
   DataGridHeadCell,
   DataGridRow,
   DataGridCell,
+  useEndlessScrollList,
 } from "@cloudoperators/juno-ui-components"
 import PeaksListItem from "./PeaksListItem"
 import HintNotFound from "../shared/HintNotFound"
 import { useGlobalsActions } from "../StoreProvider"
-import { useEndlessScrollList } from "@cloudoperators/juno-utils"
 
 const LIST_COLUMNS = 6